requirements.txt
Deprecated==1.2.13
Django==3.2.8
Pillow==8.4.0
asgiref==3.4.1
django-dotenv==1.4.2
djangorestframework==3.12.4
pip==21.3.1
postgis==1.0.4
psycopg2==2.9.1
psycopg2-binary==2.9.3
pytz==2021.3
redis==4.0.2
setuptools==58.2.0
sqlparse==0.4.2
wrapt==1.13.3